requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
psycopg2-binary>=2.9.0
ibapi>=9.76.1
requests>=2.26.0
schedule>=1.1.0
joblib>=1.1.0
ta>=0.7.0
python-dotenv>=0.19.0
tqdm>=4.62.0
pytz>=2021.1
streamlit>=1.12.0
plotly>=5.10.0
hyperopt>=0.2.7